Disable on-screen elements of incoming call and in-call UI
Sapphire (and Dream) devices have hard SEND and END keys and
users upgrading to Froyo from earlier versions may find that
the new slide-to-answer UI is unintuitive. Worse, in memory
pressure scenario the slider may not be actionable for some
time, falsely suggesting to the user that he cannot answer
an incoming call (when in fact he can press the hard key).

This change ensures that the UI for incoming calls and also
for in-call mgmt remains consistent with the Donut version.

Change-Id: Ifa24d0cb35e7f627757ab1f00c141561f91a7949